pyproject.toml
[tool.poetry]
name = "mello"
version = "0.1.0"
description = ""
authors = ["itzAyrun <ayrun3412@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord = "^2.3.2"
tomli = "^2.0.2"
tomli-w = "^1.1.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
